Handle version constraints
Terraform versions in Terraform Cloud workspaces may have version
constraints. The exporter raises an exception on version constraints.

For example, exporting a Terraform Cloud workspace set to Terraform
`~>1.5.0` raises `ValueError: ~>1.5.0 is not valid SemVer string`.

This commit proposes to add a simple utility function for parsing the
Terraform version. The function tries to remove version constraint
characters and parse the version. If any exceptions are raised, the
function will return "1.5.7" as suggested by the previous code.

https://opentofu.org/docs/language/expressions/version-constraints/